+++ src/pygeon/__init__.py
@@ -53,8 +53,7 @@ from pygeon.discretizations.fem.mat_l2 import (
     MatPwQuadratics,
 )
 
-from pygeon.discretizations.fvm.tpsa import TPSA
-from pygeon.discretizations.fvm.bcs import TPSA_BC
+from pygeon.discretizations.fvm.tpsa import TPSA, TPSA_BC
 
 from pygeon.discretizations.vem.hdiv import VRT0, VBDM1
 from pygeon.discretizations.vem.h1 import VLagrange1
